requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.47.0
httpx==0.27.2
psycopg2-binary==2.9.9
requests==2.32.3
